Warn users to make sure overriding is enabled
For a long time I didn't understand why I couldn't override anything
even though "Allow temporary override for these sites" was checked in my
Block Set options. I was honestly about to file it as a bug before I
checked the code and realized there was another place entirely where
overriding has to be enabled.

If we wanted to be less intrusive, we could have the warning only be
visible if the "allow override" option is checked. I figured it would be
better to opt for the simplest solution first though.