Adjust wildcard tip
As a new user, I tried to make a quick rule to block the reddit homepage
and subreddits (unproductive scrolling) but not reddit Google results
(useful) and tried to make it happen with just the `*` wildcard, with no
luck.

Only after trying for a while I realized there's documentation for
wildcards and it mentions the `**` wildcard which breaks through `/`
characters:

```
*.reddit.com/*
+*.reddit.com/**comments**
```

I thought it would be nice to modify the tip a bit to make this feature
more visible outside the documentation.

This is an opinionated change.